Rename simulai/regression/pytorch to simulai/regression/_pytorch
